UI & functionality updates
- floating delete button when editing a note, with confirmation
- show last update date/time when editing a note
- show keyboard upon adding/editing a note
- checks if both title/note is empty, will not proceed to add
- checks if both title/note is empty, will delete note
- removed the detail fragment
- added border per note
- converted primary key to Long
- hides the title or note if empty on recyclerView
- upon tapping a note, proceeds to editing
- convert note to multiline
- updated color & themes

+++ app/src/main/java/com/example/kotlinassessmentactivity/data/NoteDao.kt
@@ -19,7 +19,7 @@ interface NoteDao {
     fun getNotes(): Flow<List<Note>>
 
     @Query("SELECT * from note WHERE id = :id")
-    fun getNote(id: Int): Flow<Note>
+    fun getNote(id: Long): Flow<Note>
 
     // Specify the conflict strategy as IGNORE, when the user tries to add an
     // existing Note into the database Room ignores the conflict.
